Penalized maximal F (PMF) test

PMF allows the time series being tested to have a linear trend throughout the whole period of the data record (i.e., no shift in the trend component; see Wang 2003), with the annual cycle, linear trend, and lag-1 autocorrelation of the base series being estimated in tandem through iterative procedures, while accounting for all the identified mean-shifts (Wang 2008a). No reference series will be used in any of these functions; the base series is tested in this case.

FindU(
  InSeries = NULL,
  output = "./OUTPUT/example01",
  MissingValueCode = "-999.99",
  GUI = FALSE,
  plev = 0.95,
  Iadj = 10000,
  Mq = 10,
  Ny4a = 0,
  is_plot = FALSE
)

Arguments

InSeries

file path of the input data

output

prefix of the outputs

MissingValueCode

string, missing value code

GUI

boolean

plev

plev is the nominal level of confidence; (1 - plev) is the nominal level of significance, choose one of the following 6 plev values: 0.75, 0.80, 0.90, 0.95, 0.99, 0.9999

Iadj

If Iadj = 0, the data series is adjusted to the longest segment; otherwise the data series is adjusted to the chosen segment Iadj (if the given integer Iadj > Ns+1, we re-set Iadj=Ns+1, which corresponds to adjusting the series to the last segment). Set Iadj = 10000 if you want to ensure that the series is adjusted to the last segment

Mq

Mq (=0, 1, 2, ..., 20) is the number of points (categories) on which the empirical cumulative distribution function are estimated. If Mq=0, the actual Mq is determined by the length of the shortest segment Mq=1 corresponds to mean adjustments (one adjustment for all data in the same segment).

Ny4a

Value

Example1_1Cs.txt, Example1_Ustat.txt, Example1_U.dat, Example1_U.pdf

Example1_1Cs.txt

Ns >= 0 is the number of changepoints identified; changepoint positions are Ip(1), Ip(2),...,Ip(Ns)

OutFile_U.dat

index

date

original base series

estimated linear trend line of the base series

the mean-adjusted base series #'

the 6 th and 7 th columns are similar to the 4 th and 5 th columns but for shift-sizes estimated from the de-seasonalized base series

the 8 th and 9 th columns are respectively the de-seasonalized base series (i.e., the base series with its the mean annual cycle subtracted) and its multi-phase regression model fit;

the 10 th column: the estimated mean annual cycle together with the linear trend and mean-shifts;

the 11 th column: the QM-adjusted base series (the QM-adjustments here are estimated using the reference series);

the 12 th column: the multi-phase regression model fit to the de-seasonalized base series without accounting for any shifts (i.e. ignore all shifts identified).

Examples

infile = system.file("extdata/Example1.dat", package = "RHtests")
FindU(infile)
#> $fit
#> [data.table]: 
#> # A data frame: 636 × 10
#>       id date        base base_trend+m…¹ base_…² base_…³ fit_o…⁴ annua…⁵ QM_ad…⁶
#>    <dbl> <date>     <dbl>          <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1     1 1953-01-01 14.0            12.0    9.47    4.86    3.67    12.8    8.80
#>  2     2 1953-02-01 12.0            12.0    7.48    3.02    3.67    12.6    7.54
#>  3     3 1953-03-01 13.0            12.0    8.51    3.70    3.67    13.0    8.15
#>  4     4 1953-04-01 11.3            12.0    6.75    1.99    3.68    12.9    6.99
#>  5     5 1953-05-01 12.2            12.0    7.70    3.72    3.68    12.2    7.64
#>  6     6 1953-06-01 11.0            12.0    6.45    2.88    3.68    11.8    6.74
#>  7     7 1953-07-01 11.5            12.0    6.97    4.20    3.68    11.0    7.16
#>  8     8 1953-08-01  9.54           12.0    5.03    2.84    3.69    10.4    5.50
#>  9     9 1953-09-01 10.6            12.0    6.04    4.08    3.69    10.2    6.40
#> 10    10 1953-10-01  9.89           12.0    5.38    2.60    3.69    11.0    5.84
#> # … with 626 more rows, 1 more variable: fit_of_deseason_base <dbl>, and
#> #   abbreviated variable names ¹​`base_trend+meanShift`, ²​base_mean_adj,
#> #   ³​base_anomaly, ⁴​fit_of_base_anomaly, ⁵​`annualC+meanShifts`, ⁶​QM_adjusted
#> 
#> $TP
#> [data.table]: 
#> # A data frame: 3 × 9
#>    kind Idc        date probL probU  plev   PFx PFx95l PFx95h
#>   <dbl> <chr>     <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#> 1     1 "Yes " 19700100     1     1  0.95  49.2   14.3   19.2
#> 2     1 "Yes " 19740200     1     1  0.95  23.4   13.3   17.7
#> 3     1 "Yes " 19760600     1     1  0.95  21.4   14.5   19.5
#>