FindUD

FindUD

FindUD(
  InSeries = NULL,
  InCs,
  output,
  MissingValueCode = "-999.99",
  GUI = FALSE,
  plev = 0.95,
  Iadj = 10000,
  Mq = 10,
  Ny4a = 0,
  is_plot = FALSE
)

Arguments

InSeries

file path of the input data

output

prefix of the outputs

MissingValueCode

string, missing value code

GUI

boolean

plev

plev is the nominal level of confidence; (1 - plev) is the nominal level of significance, choose one of the following 6 plev values: 0.75, 0.80, 0.90, 0.95, 0.99, 0.9999

Iadj

If Iadj = 0, the data series is adjusted to the longest segment; otherwise the data series is adjusted to the chosen segment Iadj (if the given integer Iadj > Ns+1, we re-set Iadj=Ns+1, which corresponds to adjusting the series to the last segment). Set Iadj = 10000 if you want to ensure that the series is adjusted to the last segment

Mq

Mq (=0, 1, 2, ..., 20) is the number of points (categories) on which the empirical cumulative distribution function are estimated. If Mq=0, the actual Mq is determined by the length of the shortest segment Mq=1 corresponds to mean adjustments (one adjustment for all data in the same segment).

Ny4a