Penalized maximal F (PMF) test

PMF allows the time series being tested to have a linear trend throughout the whole period of the data record (i.e., no shift in the trend component; see Wang 2003), with the annual cycle, linear trend, and lag-1 autocorrelation of the base series being estimated in tandem through iterative procedures, while accounting for all the identified mean-shifts (Wang 2008a). No reference series will be used in any of these functions; the base series is tested in this case.

FindU(
  InSeries = NULL,
  output = "./OUTPUT/example01",
  MissingValueCode = "-999.99",
  GUI = FALSE,
  plev = 0.95,
  Iadj = 10000,
  Mq = 10,
  Ny4a = 0,
  is_plot = FALSE
)

Arguments

InSeries

file path of the input data

output

prefix of the outputs

MissingValueCode

string, missing value code

GUI

boolean

plev

plev is the nominal level of confidence; (1 - plev) is the nominal level of significance, choose one of the following 6 plev values: 0.75, 0.80, 0.90, 0.95, 0.99, 0.9999

Iadj

If Iadj = 0, the data series is adjusted to the longest segment; otherwise the data series is adjusted to the chosen segment Iadj (if the given integer Iadj > Ns+1, we re-set Iadj=Ns+1, which corresponds to adjusting the series to the last segment). Set Iadj = 10000 if you want to ensure that the series is adjusted to the last segment

Mq

Mq (=0, 1, 2, ..., 20) is the number of points (categories) on which the empirical cumulative distribution function are estimated. If Mq=0, the actual Mq is determined by the length of the shortest segment Mq=1 corresponds to mean adjustments (one adjustment for all data in the same segment).

Ny4a

Value

Example1_1Cs.txt, Example1_Ustat.txt, Example1_U.dat, Example1_U.pdf

Example1_1Cs.txt

Ns >= 0 is the number of changepoints identified; changepoint positions are Ip(1), Ip(2),...,Ip(Ns)

OutFile_U.dat

index

date

original base series

estimated linear trend line of the base series

the mean-adjusted base series #'

the 6 th and 7 th columns are similar to the 4 th and 5 th columns but for shift-sizes estimated from the de-seasonalized base series

the 8 th and 9 th columns are respectively the de-seasonalized base series (i.e., the base series with its the mean annual cycle subtracted) and its multi-phase regression model fit;

the 10 th column: the estimated mean annual cycle together with the linear trend and mean-shifts;

the 11 th column: the QM-adjusted base series (the QM-adjustments here are estimated using the reference series);

the 12 th column: the multi-phase regression model fit to the de-seasonalized base series without accounting for any shifts (i.e. ignore all shifts identified).

Examples

infile = system.file("extdata/Example1.dat", package = "RHtests")
FindU(infile)
#> Warning: cannot create file './OUTPUT/example01_U.dat', reason 'No such file or directory'
#> Warning: cannot create file './OUTPUT/example01_U.pdf', reason 'No such file or directory'
#> Warning: cannot create file './OUTPUT/example01_Ustat.txt', reason 'No such file or directory'
#> Warning: cannot create file './OUTPUT/example01_1Cs.txt', reason 'No such file or directory'
#> Warning: cannot open file './OUTPUT/example01_Ustat.txt': No such file or directory
#> Error in file(file, ifelse(append, "a", "w")): cannot open the connection